change int64 to int32, so that the algorithm could work on gpu

+++ pylearn2/datasets/csv_dataset.py
@@ -125,7 +125,7 @@ class CSVDataset(DenseDesignMatrix):
         if self.task == 'regression':
             super(CSVDataset, self).__init__(X=X, y=y, **kwargs)
         else:
-            y = y.astype('int64')
+            y = y.astype('int32')
             super(CSVDataset, self).__init__(X=X, y=y,
                                              y_labels=np.max(y) + 1, **kwargs)
